Filled in almost every field required for pb to geth block conversion
