+++ cmd/geth/chaincmd.go
@@ -26,7 +26,6 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/encoding/gzip"
 	"io"
-	"math/big"
 	"os"
 	"path/filepath"
 	"regexp"
@@ -55,8 +54,6 @@ import (
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/node"
 	"github.com/ethereum/go-ethereum/params"
-	"github.com/ethereum/go-ethereum/rlp"
-	"github.com/ethereum/go-ethereum/trie"
 	"github.com/streamingfast/firehose-core/firehose/client"
 	"github.com/urfave/cli/v2"
 )
@@ -990,286 +987,4 @@ func exportFromFirehose(ctx *cli.Context) error {
 		fmt.Println("Export completed successfully.")
 		return nil
 	}
-}
-
-// convertFirehoseBlockToGethBlock converts a Firehose protobuf block to a geth Block
-func convertFirehoseBlockToGethBlock(pbBlock *pbeth.Block) (*types.Block, error) {
-	if pbBlock == nil || pbBlock.Header == nil {
-		return nil, fmt.Errorf("invalid block or header")
-	}
-
-	// Convert header
-	header := &types.Header{
-		ParentHash:  common.Hash(pbBlock.Header.ParentHash),
-		UncleHash:   common.Hash(pbBlock.Header.UncleHash),
-		Coinbase:    common.Address(pbBlock.Header.Coinbase),
-		Root:        common.Hash(pbBlock.Header.StateRoot),
-		TxHash:      common.Hash(pbBlock.Header.TransactionsRoot),
-		ReceiptHash: common.Hash(pbBlock.Header.ReceiptRoot),
-		Bloom:       types.BytesToBloom(pbBlock.Header.LogsBloom),
-		Difficulty:  pbBlock.Header.Difficulty.Native(),
-		Number:      big.NewInt(int64(pbBlock.Header.Number)),
-		GasLimit:    pbBlock.Header.GasLimit,
-		GasUsed:     pbBlock.Header.GasUsed,
-		Time:        uint64(pbBlock.Header.Timestamp.Seconds),
-		Extra:       pbBlock.Header.ExtraData,
-		MixDigest:   common.Hash(pbBlock.Header.MixHash),
-		Nonce:       types.EncodeNonce(pbBlock.Header.Nonce),
-	}
-
-	if pbBlock.Header.BaseFeePerGas != nil {
-		header.BaseFee = pbBlock.Header.BaseFeePerGas.Native()
-	}
-
-	if pbBlock.Header.WithdrawalsRoot != nil {
-		header.WithdrawalsHash = (*common.Hash)(pbBlock.Header.WithdrawalsRoot)
-	}
-
-	if pbBlock.Header.BlobGasUsed != nil {
-		header.BlobGasUsed = pbBlock.Header.BlobGasUsed
-	}
-
-	if pbBlock.Header.ExcessBlobGas != nil {
-		header.ExcessBlobGas = pbBlock.Header.ExcessBlobGas
-	}
-
-	if pbBlock.Header.ParentBeaconRoot != nil {
-		header.ParentBeaconRoot = (*common.Hash)(pbBlock.Header.ParentBeaconRoot)
-	}
-
-	if pbBlock.Header.RequestsHash != nil {
-		header.ParentBeaconRoot = (*common.Hash)(pbBlock.Header.RequestsHash)
-	}
-
-	// Convert transactions
-	var txs []*types.Transaction
-	var receipts types.Receipts
-	for _, pbTx := range pbBlock.TransactionTraces {
-		if pbTx == nil {
-			continue
-		}
-
-		var tx *types.Transaction
-		switch pbTx.Type {
-		case 0: // LegacyTx
-			tx = types.NewTx(&types.LegacyTx{
-				Nonce:    pbTx.Nonce,
-				GasPrice: pbTx.GasPrice.Native(),
-				Gas:      pbTx.GasLimit,
-				To:       func() *common.Address { addr := common.BytesToAddress(pbTx.To); return &addr }(),
-				Value:    pbTx.Value.Native(),
-				Data:     pbTx.Input,
-				V:        new(big.Int).SetBytes(pbTx.V),
-				R:        new(big.Int).SetBytes(pbTx.R),
-				S:        new(big.Int).SetBytes(pbTx.S),
-			})
-		case 1: // AccessListTx
-			tx = types.NewTx(&types.AccessListTx{
-				ChainID:    extractChainIDFromSetCodeAuth(pbTx.SetCodeAuthorizations).ToBig(),
-				Nonce:      pbTx.Nonce,
-				GasPrice:   pbTx.GasPrice.Native(),
-				Gas:        pbTx.GasLimit,
-				To:         func() *common.Address { addr := common.BytesToAddress(pbTx.To); return &addr }(),
-				Value:      pbTx.Value.Native(),
-				Data:       pbTx.Input,
-				AccessList: convertFirehoseAccessList(pbTx.AccessList),
-				V:          new(big.Int).SetBytes(pbTx.V),
-				R:          new(big.Int).SetBytes(pbTx.R),
-				S:          new(big.Int).SetBytes(pbTx.S),
-			})
-		case 2: // DynamicFeeTx
-			tx = types.NewTx(&types.DynamicFeeTx{
-				ChainID:    extractChainIDFromSetCodeAuth(pbTx.SetCodeAuthorizations).ToBig(),
-				Nonce:      pbTx.Nonce,
-				GasTipCap:  pbTx.MaxPriorityFeePerGas.Native(),
-				GasFeeCap:  pbTx.MaxFeePerGas.Native(),
-				Gas:        pbTx.GasLimit,
-				To:         func() *common.Address { addr := common.BytesToAddress(pbTx.To); return &addr }(),
-				Value:      pbTx.Value.Native(),
-				Data:       pbTx.Input,
-				AccessList: convertFirehoseAccessList(pbTx.AccessList),
-				V:          new(big.Int).SetBytes(pbTx.V),
-				R:          new(big.Int).SetBytes(pbTx.R),
-				S:          new(big.Int).SetBytes(pbTx.S),
-			})
-		case 3: // BlobTx
-			tx = types.NewTx(&types.BlobTx{
-				ChainID:    extractChainIDFromSetCodeAuth(pbTx.SetCodeAuthorizations),
-				Nonce:      pbTx.Nonce,
-				GasTipCap:  bigIntToUint256(pbTx.MaxPriorityFeePerGas.Native()),
-				GasFeeCap:  bigIntToUint256(pbTx.MaxFeePerGas.Native()),
-				Gas:        pbTx.GasLimit,
-				To:         common.BytesToAddress(pbTx.To),
-				Value:      bigIntToUint256(pbTx.Value.Native()),
-				Data:       pbTx.Input,
-				AccessList: convertFirehoseAccessList(pbTx.AccessList),
-				BlobFeeCap: bigIntToUint256(pbTx.BlobGasFeeCap.Native()),
-				BlobHashes: convertFirehoseBlobHashes(pbTx.BlobHashes),
-				V:          bigIntToUint256(new(big.Int).SetBytes(pbTx.V)),
-				R:          bigIntToUint256(new(big.Int).SetBytes(pbTx.R)),
-				S:          bigIntToUint256(new(big.Int).SetBytes(pbTx.S)),
-			})
-		case 4: // SetCodeTx
-			tx = types.NewTx(&types.SetCodeTx{
-				ChainID:    extractChainIDFromSetCodeAuth(pbTx.SetCodeAuthorizations),
-				Nonce:      pbTx.Nonce,
-				GasTipCap:  bigIntToUint256(pbTx.MaxPriorityFeePerGas.Native()),
-				GasFeeCap:  bigIntToUint256(pbTx.MaxFeePerGas.Native()),
-				Gas:        pbTx.GasLimit,
-				To:         common.BytesToAddress(pbTx.To),
-				Value:      bigIntToUint256(pbTx.Value.Native()),
-				Data:       pbTx.Input,
-				AccessList: convertFirehoseAccessList(pbTx.AccessList),
-				AuthList:   convertFirehoseSetCodeAuthorizations(pbTx.SetCodeAuthorizations),
-				V:          bigIntToUint256(new(big.Int).SetBytes(pbTx.V)),
-				R:          bigIntToUint256(new(big.Int).SetBytes(pbTx.R)),
-				S:          bigIntToUint256(new(big.Int).SetBytes(pbTx.S)),
-			})
-		default:
-			continue
-		}
-
-		if tx != nil {
-			txs = append(txs, tx)
-		}
-
-		// Convert receipt if present
-		if pbTx.Receipt != nil {
-			receipt := &types.Receipt{
-				Type:              uint8(pbTx.Type),
-				PostState:         pbTx.Receipt.StateRoot,
-				CumulativeGasUsed: pbTx.Receipt.CumulativeGasUsed,
-				Bloom:             types.BytesToBloom(pbTx.Receipt.LogsBloom),
-				Logs:              convertFirehoseLogsToGethLogs(pbTx.Receipt.Logs),
-
-				TxHash:  common.Hash(pbTx.Hash),
-				GasUsed: pbTx.GasUsed,
-			}
-
-			// Status
-			if len(pbTx.Receipt.StateRoot) == 0 {
-				if pbTx.Status == 1 {
-					receipt.Status = types.ReceiptStatusSuccessful
-				} else {
-					receipt.Status = types.ReceiptStatusFailed
-				}
-			}
-
-			// ContractAddress
-			if pbTx.To == nil || len(pbTx.To) == 0 {
-				// TODO
-				receipt.ContractAddress = common.Address{}
-			}
-
-			// EffectiveGasPrice
-			if pbTx.GasPrice != nil {
-				// TODO
-				receipt.EffectiveGasPrice = pbTx.GasPrice.Native()
-			}
-
-			// BlobGasUsed
-			if pbTx.Receipt.BlobGasUsed != nil {
-				receipt.BlobGasUsed = *pbTx.Receipt.BlobGasUsed
-			}
-			if pbTx.Receipt.BlobGasPrice != nil {
-				receipt.BlobGasPrice = pbTx.Receipt.BlobGasPrice.Native()
-			}
-
-			// BlockHash, BlockNumber, TransactionIndex: not available in pbTx, set to zero/nil
-			// These will be derived later by Receipts.DeriveFields
-			// TODO
-			receipt.BlockHash = common.Hash{}
-			receipt.BlockNumber = nil
-			receipt.TransactionIndex = 0
-			receipts = append(receipts, receipt)
-		}
-	}
-
-	// Convert uncles
-	var uncles []*types.Header
-	for _, pbUncle := range pbBlock.Uncles {
-		if pbUncle == nil {
-			continue
-		}
-		uncle := &types.Header{
-			ParentHash:  common.Hash(pbUncle.ParentHash),
-			UncleHash:   common.Hash(pbUncle.UncleHash),
-			Coinbase:    common.Address(pbUncle.Coinbase),
-			Root:        common.Hash(pbUncle.StateRoot),
-			TxHash:      common.Hash(pbUncle.TransactionsRoot),
-			ReceiptHash: common.Hash(pbUncle.ReceiptRoot),
-			Bloom:       types.BytesToBloom(pbUncle.LogsBloom),
-			Difficulty:  pbUncle.Difficulty.Native(),
-			Number:      big.NewInt(int64(pbUncle.Number)),
-			GasLimit:    pbUncle.GasLimit,
-			GasUsed:     pbUncle.GasUsed,
-			Time:        uint64(pbUncle.Timestamp.Seconds),
-			Extra:       pbUncle.ExtraData,
-			MixDigest:   common.Hash(pbUncle.MixHash),
-			Nonce:       types.EncodeNonce(pbUncle.Nonce),
-		}
-		if pbUncle.BaseFeePerGas != nil {
-			uncle.BaseFee = pbUncle.BaseFeePerGas.Native()
-		}
-
-		if pbUncle.WithdrawalsRoot != nil {
-			uncle.WithdrawalsHash = (*common.Hash)(pbUncle.WithdrawalsRoot)
-		}
-
-		if pbUncle.BlobGasUsed != nil {
-			uncle.BlobGasUsed = pbUncle.BlobGasUsed
-		}
-
-		if pbUncle.ExcessBlobGas != nil {
-			uncle.ExcessBlobGas = pbUncle.ExcessBlobGas
-		}
-
-		if pbUncle.ParentBeaconRoot != nil {
-			uncle.ParentBeaconRoot = (*common.Hash)(pbUncle.ParentBeaconRoot)
-		}
-
-		if pbUncle.RequestsHash != nil {
-			uncle.ParentBeaconRoot = (*common.Hash)(pbUncle.RequestsHash)
-		}
-
-		uncles = append(uncles, uncle)
-	}
-
-	body := &types.Body{
-		Transactions: txs,
-		Uncles:       uncles,
-		// TODO: Withdrawal
-	}
-
-	return types.NewBlock(header, body, receipts, trie.NewStackTrie(nil)), nil
-}
-
-// writeBatch writes a batch of blocks to an RLP file
-func writeBatch(blocks []*types.Block, outputPrefix string, batchNum int) error {
-	dir := "rlp-data"
-	if err := os.MkdirAll(dir, 0o755); err != nil {
-		return fmt.Errorf("failed to create directory %s: %w", dir, err)
-	}
-
-	var filename string
-	if batchNum == 0 {
-		filename = filepath.Join(dir, fmt.Sprintf("%s.rlp", outputPrefix))
-	} else {
-		filename = filepath.Join(dir, fmt.Sprintf("%s.rlp.%d", outputPrefix, batchNum))
-	}
-
-	file, err := os.Create(filename)
-	if err != nil {
-		return fmt.Errorf("failed to create file %s: %w", filename, err)
-	}
-	defer file.Close()
-
-	// Write each block as RLP encoded data
-	for _, block := range blocks {
-		if err := rlp.Encode(file, block); err != nil {
-			return fmt.Errorf("failed to encode block %d: %w", block.NumberU64(), err)
-		}
-	}
-
-	return nil
 }
